+++ packages/contentful--app-scripts/src/clean-up/get-clean-up-settings.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import ora from 'ora';
-import { DEFAULT_BUNDLES_TO_KEEP } from '../../utils/constants';
+import { DEFAULT_BUNDLES_TO_KEEP } from '../constants';
 import { validateArguments } from '../validate-arguments';
 import { getAppInfo } from '../get-app-info';
 import { CleanupOptions, CleanupSettings } from '.';
