+++ packages/contentful--app-scripts/src/constants.ts
