+++ packages/contentful--app-scripts/src/open/open-settings.ts
@@ -1,7 +1,7 @@
 import open from 'open';
 import chalk from 'chalk';
 import inquirer from 'inquirer';
-import { APP_DEF_ENV_KEY } from '../../utils/constants';
+import { APP_DEF_ENV_KEY } from '../constants';
 import { OpenSettingsOptions } from '.';
 
 export const REDIRECT_URL = 'https://app.contentful.com/deeplink?link=app-definition';
